feat: Implement expenses editing. style: Improve UI responsiveess for mobile sites when adding or editing expenses.
