+++ back-end/expense/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from .models import Expense, SplitDetail
 
 class ExpenseAdmin(admin.ModelAdmin):
-    list_display = ['description', 'paid_by', 'total_amount', 'created_on']
+    list_display = ['id', 'description', 'paid_by', 'total_amount', 'created_on']
 
 admin.site.register(Expense, ExpenseAdmin)
