+++ front-end/src/components/expense/SplitDetailsDialog.jsx
@@ -2,23 +2,15 @@ import { CustomDialogHeader } from '../generic/styles/CustomDIalogHeader';
 // prettier-ignore
 import { equallyCalculatedValue, unequallyCalculatedValues, } from './functions/SplitValueCalculations';
 // prettier-ignore
-import { IconButton, DialogContent, Typography, Grid, Paper, Tabs, TextField, InputAdornment, Divider, Checkbox } from '@mui/material';
+import { IconButton, Dialog, DialogContent, Typography, Grid, Paper, Tabs, TextField, InputAdornment, Divider, Checkbox } from '@mui/material';
 import { StyledTab, StyledTabPanel } from './styles/SplitDetailsTabStyles';
 import { useState, useEffect } from 'react';
 import CloseIcon from '@mui/icons-material/Close';
 import EqualizerIcon from '@mui/icons-material/Equalizer';
 
 const SplitDetailsDialog = (props) => {
-    const {
-        open,
-        setOpen,
-        splitDetail,
-        setSplitDetail,
-        paidBy,
-        totalAmount,
-        splitBetweenUsers,
-        expenseType,
-    } = props;
+    const { open, setOpen, splitDetail, setSplitDetail, paidBy, totalAmount, splitBetweenUsers } =
+        props;
 
     const [splitMethod, setSplitType] = useState(splitDetail?.splitMethod || 'E');
 
@@ -123,125 +115,138 @@ const SplitDetailsDialog = (props) => {
     const calculatedTotalText = calculateTotalTxtValue();
 
     return (
-        <Grid item xs={12} md={5} sx={{ display: 'flex', justifyContent: 'center', p: 1 }}>
-            <Paper elevation={2} style={{ maxWidth: '99dvw' }}>
-                <CustomDialogHeader item>
-                    Split Details
-                    <IconButton size='small' onClick={() => setOpen(false)}>
-                        <CloseIcon sx={{ color: 'white' }} />
-                    </IconButton>
-                </CustomDialogHeader>
-                <DialogContent dividers sx={{ padding: '5px' }}>
-                    <Grid container direction='row' justify='flex-end' alignItems='center'>
-                        <Grid item xs={12}>
-                            <Tabs
-                                value={splitMethod}
-                                onChange={handleChange}
-                                variant='fullWidth'
-                                TabIndicatorProps={{
-                                    style: { backgroundColor: 'transparent' },
-                                }}
-                            >
-                                <StyledTab value='E' label='=' tooltipText='Split Equally' />
-                                <StyledTab value='P' label='%' tooltipText='Split in Percentage' />
-                                <StyledTab
-                                    value='U'
-                                    label='1.23'
-                                    tooltipText='Split in Exact Amounts'
-                                />
-                                <StyledTab
-                                    value='S'
-                                    icon={<EqualizerIcon />}
-                                    tooltipText='Split in Shares'
-                                />
-                            </Tabs>
-                        </Grid>
-                    </Grid>
-                    <StyledTabPanel value={splitMethod} index={'E'}>
-                        {splitMethod === 'E' && (
+        <Dialog disableScrollLock={true} open={open} onClose={() => {}}>
+            <Grid item xs={12} sx={{ display: 'flex', justifyContent: 'center', p: 1 }}>
+                <Paper elevation={0}>
+                    <CustomDialogHeader item>
+                        Split Details
+                        <IconButton size='small' onClick={() => setOpen(false)}>
+                            <CloseIcon sx={{ color: 'white' }} />
+                        </IconButton>
+                    </CustomDialogHeader>
+                    <DialogContent dividers sx={{ padding: '5px' }}>
+                        <Grid container direction='row' justify='flex-end' alignItems='center'>
                             <Grid item xs={12}>
-                                <Grid container>
-                                    <Grid item xs={12}>
-                                        <Typography variant='h6' sx={{ fontWeight: 600 }}>
-                                            Split Equally
-                                        </Typography>
-                                    </Grid>
-                                    <Grid item xs={12}>
-                                        {splitDetail?.splitValues?.map((infoDict) => (
-                                            <Grid container key={infoDict.user.email}>
-                                                <Grid
-                                                    item
-                                                    xs={1}
-                                                    sx={{ display: 'flex', alignItems: 'start' }}
-                                                >
-                                                    <Checkbox
-                                                        defaultChecked
-                                                        sx={{ p: 0 }}
-                                                        id={`included-${infoDict.user?.id}`}
-                                                        onChange={handleIncludeChange}
-                                                    />
+                                <Tabs
+                                    value={splitMethod}
+                                    onChange={handleChange}
+                                    variant='fullWidth'
+                                    TabIndicatorProps={{
+                                        style: { backgroundColor: 'transparent' },
+                                    }}
+                                >
+                                    <StyledTab value='E' label='=' tooltipText='Split Equally' />
+                                    <StyledTab
+                                        value='P'
+                                        label='%'
+                                        tooltipText='Split in Percentage'
+                                    />
+                                    <StyledTab
+                                        value='U'
+                                        label='1.23'
+                                        tooltipText='Split in Exact Amounts'
+                                    />
+                                    <StyledTab
+                                        value='S'
+                                        icon={<EqualizerIcon />}
+                                        tooltipText='Split in Shares'
+                                    />
+                                </Tabs>
+                            </Grid>
+                        </Grid>
+                        <StyledTabPanel value={splitMethod} index={'E'}>
+                            {splitMethod === 'E' && (
+                                <Grid item xs={12}>
+                                    <Grid container>
+                                        <Grid item xs={12}>
+                                            <Typography variant='h6' sx={{ fontWeight: 600 }}>
+                                                Split Equally
+                                            </Typography>
+                                        </Grid>
+                                        <Grid item xs={12}>
+                                            {splitDetail?.splitValues?.map((infoDict) => (
+                                                <Grid container key={infoDict.user.email}>
+                                                    <Grid
+                                                        item
+                                                        xs={1}
+                                                        sx={{
+                                                            display: 'flex',
+                                                            alignItems: 'start',
+                                                        }}
+                                                    >
+                                                        <Checkbox
+                                                            defaultChecked
+                                                            sx={{ p: 0 }}
+                                                            id={`included-${infoDict.user?.id}`}
+                                                            onChange={handleIncludeChange}
+                                                        />
+                                                    </Grid>
+                                                    <Grid item xs={9}>
+                                                        {infoDict.user.display_name}
+                                                    </Grid>
+                                                    <Grid item xs={2} sx={{ textAlign: 'end' }}>
+                                                        $
+                                                        {Math.abs(
+                                                            infoDict.calculatedAmount
+                                                        ).toFixed(2)}
+                                                    </Grid>
                                                 </Grid>
-                                                <Grid item xs={9}>
-                                                    {infoDict.user.display_name}
+                                            ))}
+                                        </Grid>
+                                        <Grid item xs={12}>
+                                            <Grid container>
+                                                <Grid item xs={8}>
+                                                    <strong>Total:</strong>
                                                 </Grid>
-                                                <Grid item xs={2} sx={{ textAlign: 'end' }}>
-                                                    $
-                                                    {Math.abs(infoDict.calculatedAmount).toFixed(2)}
+                                                <Grid item xs={4} sx={{ textAlign: 'end' }}>
+                                                    <strong>${totalAmount}</strong>
                                                 </Grid>
-                                            </Grid>
-                                        ))}
-                                    </Grid>
-                                    <Grid item xs={12}>
-                                        <Grid container>
-                                            <Grid item xs={8}>
-                                                <strong>Total:</strong>
-                                            </Grid>
-                                            <Grid item xs={4} sx={{ textAlign: 'end' }}>
-                                                <strong>${totalAmount}</strong>
                                             </Grid>
                                         </Grid>
                                     </Grid>
                                 </Grid>
-                            </Grid>
-                        )}
-                    </StyledTabPanel>
-                    <StyledTabPanel value={splitMethod} index={'P'}>
-                        {splitMethod === 'P' && (
-                            <Grid item xs={12}>
-                                <Grid container>
-                                    <Grid item xs={12}>
-                                        <Typography variant='h6' sx={{ fontWeight: 600 }}>
-                                            Split in Percentage
-                                        </Typography>
-                                    </Grid>
-                                    <Grid item xs={12}>
-                                        {splitDetail?.splitValues?.map((infoDict) => (
-                                            <Grid
-                                                container
-                                                key={infoDict.user.email}
-                                                sx={{ mt: 1 }}
-                                            >
+                            )}
+                        </StyledTabPanel>
+                        <StyledTabPanel value={splitMethod} index={'P'}>
+                            {splitMethod === 'P' && (
+                                <Grid item xs={12}>
+                                    <Grid container>
+                                        <Grid item xs={12}>
+                                            <Typography variant='h6' sx={{ fontWeight: 600 }}>
+                                                Split in Percentage
+                                            </Typography>
+                                        </Grid>
+                                        <Grid item xs={12}>
+                                            {splitDetail?.splitValues?.map((infoDict) => (
                                                 <Grid
-                                                    item
-                                                    xs={1}
-                                                    sx={{ display: 'flex', alignItems: 'start' }}
+                                                    container
+                                                    key={infoDict.user.email}
+                                                    sx={{ mt: 1 }}
                                                 >
-                                                    <Checkbox
-                                                        defaultChecked
-                                                        sx={{ p: 0 }}
-                                                        id={`included-${infoDict.user?.id}`}
-                                                        onChange={handleIncludeChange}
-                                                    />
-                                                </Grid>
-                                                <Grid item xs={6}>
-                                                    {infoDict.user.display_name}
-                                                </Grid>
-                                                <Grid item xs={2} sx={{ textAlign: 'start' }}>
-                                                    {`$ ${getUsersAmt(infoDict.user)}`}
-                                                </Grid>
-                                                <Grid item xs={3} sx={{ textAlign: 'end' }}>
-                                                    {/* prettier-ignore */}
-                                                    <TextField
+                                                    <Grid
+                                                        item
+                                                        xs={1}
+                                                        sx={{
+                                                            display: 'flex',
+                                                            alignItems: 'start',
+                                                        }}
+                                                    >
+                                                        <Checkbox
+                                                            defaultChecked
+                                                            sx={{ p: 0 }}
+                                                            id={`included-${infoDict.user?.id}`}
+                                                            onChange={handleIncludeChange}
+                                                        />
+                                                    </Grid>
+                                                    <Grid item xs={6}>
+                                                        {infoDict.user.display_name}
+                                                    </Grid>
+                                                    <Grid item xs={2} sx={{ textAlign: 'start' }}>
+                                                        {`$ ${getUsersAmt(infoDict.user)}`}
+                                                    </Grid>
+                                                    <Grid item xs={3} sx={{ textAlign: 'end' }}>
+                                                        {/* prettier-ignore */}
+                                                        <TextField
                                                     id={`split-value-${infoDict.user?.id}`}
                                                     size='small'
                                                     variant='standard'
@@ -252,152 +257,159 @@ const SplitDetailsDialog = (props) => {
                                                         startAdornment: <InputAdornment position='start'>%</InputAdornment>,
                                                       }}
                                                 />
+                                                    </Grid>
                                                 </Grid>
-                                            </Grid>
-                                        ))}
-                                    </Grid>
-                                    <Grid item xs={12} sx={{ mt: 2 }}>
-                                        <Divider />
-                                        <Grid container sx={{ mt: 1 }}>
-                                            <Grid item xs={8}>
-                                                <strong>Total:</strong>
-                                            </Grid>
-                                            <Grid item xs={4} sx={{ textAlign: 'end' }}>
-                                                <strong>${calculatedTotalText}</strong>
+                                            ))}
+                                        </Grid>
+                                        <Grid item xs={12} sx={{ mt: 2 }}>
+                                            <Divider />
+                                            <Grid container sx={{ mt: 1 }}>
+                                                <Grid item xs={8}>
+                                                    <strong>Total:</strong>
+                                                </Grid>
+                                                <Grid item xs={4} sx={{ textAlign: 'end' }}>
+                                                    <strong>${calculatedTotalText}</strong>
+                                                </Grid>
                                             </Grid>
                                         </Grid>
                                     </Grid>
                                 </Grid>
-                            </Grid>
-                        )}
-                    </StyledTabPanel>
-                    <StyledTabPanel value={splitMethod} index={'U'}>
-                        {splitMethod === 'U' && (
-                            <Grid item xs={12}>
-                                <Grid container>
-                                    <Grid item xs={12}>
-                                        <Typography variant='h6' sx={{ fontWeight: 600 }}>
-                                            Split by unequal amounts
-                                        </Typography>
-                                    </Grid>
-                                    <Grid item xs={12}>
-                                        {splitDetail?.splitValues?.map((infoDict) => (
-                                            <Grid
-                                                container
-                                                key={infoDict.user.email}
-                                                sx={{ mt: 1 }}
-                                            >
+                            )}
+                        </StyledTabPanel>
+                        <StyledTabPanel value={splitMethod} index={'U'}>
+                            {splitMethod === 'U' && (
+                                <Grid item xs={12}>
+                                    <Grid container>
+                                        <Grid item xs={12}>
+                                            <Typography variant='h6' sx={{ fontWeight: 600 }}>
+                                                Split by unequal amounts
+                                            </Typography>
+                                        </Grid>
+                                        <Grid item xs={12}>
+                                            {splitDetail?.splitValues?.map((infoDict) => (
                                                 <Grid
-                                                    item
-                                                    xs={1}
-                                                    sx={{ display: 'flex', alignItems: 'start' }}
+                                                    container
+                                                    key={infoDict.user.email}
+                                                    sx={{ mt: 1 }}
                                                 >
-                                                    <Checkbox
-                                                        defaultChecked
-                                                        sx={{ p: 0 }}
-                                                        id={`included-${infoDict.user?.id}`}
-                                                        onChange={handleIncludeChange}
-                                                    />
+                                                    <Grid
+                                                        item
+                                                        xs={1}
+                                                        sx={{
+                                                            display: 'flex',
+                                                            alignItems: 'start',
+                                                        }}
+                                                    >
+                                                        <Checkbox
+                                                            defaultChecked
+                                                            sx={{ p: 0 }}
+                                                            id={`included-${infoDict.user?.id}`}
+                                                            onChange={handleIncludeChange}
+                                                        />
+                                                    </Grid>
+                                                    <Grid item xs={7}>
+                                                        {infoDict.user.display_name}
+                                                    </Grid>
+                                                    <Grid item xs={2} sx={{ textAlign: 'start' }}>
+                                                        {`$ ${getUsersAmt(infoDict.user)}`}
+                                                    </Grid>
+                                                    <Grid item xs={2} sx={{ textAlign: 'end' }}>
+                                                        <TextField
+                                                            id={`split-value-${infoDict.user?.id}`}
+                                                            size='small'
+                                                            variant='standard'
+                                                            type='number'
+                                                            value={infoDict.value || ''}
+                                                            onChange={handleTextValueChange}
+                                                        />
+                                                    </Grid>
                                                 </Grid>
-                                                <Grid item xs={7}>
-                                                    {infoDict.user.display_name}
+                                            ))}
+                                        </Grid>
+                                        <Grid item xs={12} sx={{ mt: 2 }}>
+                                            <Divider />
+                                            <Grid container sx={{ mt: 1 }}>
+                                                <Grid item xs={8}>
+                                                    <strong>Total:</strong>
                                                 </Grid>
-                                                <Grid item xs={2} sx={{ textAlign: 'start' }}>
-                                                    {`$ ${getUsersAmt(infoDict.user)}`}
+                                                <Grid item xs={4} sx={{ textAlign: 'end' }}>
+                                                    <strong>${calculatedTotalText}</strong>
                                                 </Grid>
-                                                <Grid item xs={2} sx={{ textAlign: 'end' }}>
-                                                    <TextField
-                                                        id={`split-value-${infoDict.user?.id}`}
-                                                        size='small'
-                                                        variant='standard'
-                                                        type='number'
-                                                        value={infoDict.value || ''}
-                                                        onChange={handleTextValueChange}
-                                                    />
-                                                </Grid>
-                                            </Grid>
-                                        ))}
-                                    </Grid>
-                                    <Grid item xs={12} sx={{ mt: 2 }}>
-                                        <Divider />
-                                        <Grid container sx={{ mt: 1 }}>
-                                            <Grid item xs={8}>
-                                                <strong>Total:</strong>
-                                            </Grid>
-                                            <Grid item xs={4} sx={{ textAlign: 'end' }}>
-                                                <strong>${calculatedTotalText}</strong>
                                             </Grid>
                                         </Grid>
                                     </Grid>
                                 </Grid>
-                            </Grid>
-                        )}
-                    </StyledTabPanel>
-                    <StyledTabPanel value={splitMethod} index={'S'}>
-                        {splitMethod === 'S' && (
-                            <Grid item xs={12}>
-                                <Grid container>
-                                    <Grid item xs={12}>
-                                        <Typography variant='h6' sx={{ fontWeight: 600 }}>
-                                            Split in Shares
-                                        </Typography>
-                                    </Grid>
-                                    <Grid item xs={12}>
-                                        {splitDetail?.splitValues?.map((infoDict) => (
-                                            <Grid
-                                                container
-                                                key={infoDict.user.email}
-                                                sx={{ mt: 1 }}
-                                            >
+                            )}
+                        </StyledTabPanel>
+                        <StyledTabPanel value={splitMethod} index={'S'}>
+                            {splitMethod === 'S' && (
+                                <Grid item xs={12}>
+                                    <Grid container>
+                                        <Grid item xs={12}>
+                                            <Typography variant='h6' sx={{ fontWeight: 600 }}>
+                                                Split in Shares
+                                            </Typography>
+                                        </Grid>
+                                        <Grid item xs={12}>
+                                            {splitDetail?.splitValues?.map((infoDict) => (
                                                 <Grid
-                                                    item
-                                                    xs={1}
-                                                    sx={{ display: 'flex', alignItems: 'start' }}
+                                                    container
+                                                    key={infoDict.user.email}
+                                                    sx={{ mt: 1 }}
                                                 >
-                                                    <Checkbox
-                                                        defaultChecked
-                                                        sx={{ p: 0 }}
-                                                        id={`included-${infoDict.user?.id}`}
-                                                        onChange={handleIncludeChange}
-                                                    />
+                                                    <Grid
+                                                        item
+                                                        xs={1}
+                                                        sx={{
+                                                            display: 'flex',
+                                                            alignItems: 'start',
+                                                        }}
+                                                    >
+                                                        <Checkbox
+                                                            defaultChecked
+                                                            sx={{ p: 0 }}
+                                                            id={`included-${infoDict.user?.id}`}
+                                                            onChange={handleIncludeChange}
+                                                        />
+                                                    </Grid>
+                                                    <Grid item xs={7}>
+                                                        {infoDict.user.display_name}
+                                                    </Grid>
+                                                    <Grid item xs={2} sx={{ textAlign: 'start' }}>
+                                                        {`$ ${getUsersAmt(infoDict.user)}`}
+                                                    </Grid>
+                                                    <Grid item xs={2} sx={{ textAlign: 'end' }}>
+                                                        <TextField
+                                                            id={`split-value-${infoDict.user?.id}`}
+                                                            size='small'
+                                                            variant='standard'
+                                                            type='number'
+                                                            value={infoDict.value || ''}
+                                                            onChange={handleTextValueChange}
+                                                        />
+                                                    </Grid>
                                                 </Grid>
-                                                <Grid item xs={7}>
-                                                    {infoDict.user.display_name}
+                                            ))}
+                                        </Grid>
+                                        <Grid item xs={12} sx={{ mt: 2 }}>
+                                            <Divider />
+                                            <Grid container sx={{ mt: 1 }}>
+                                                <Grid item xs={8}>
+                                                    <strong>Total:</strong>
                                                 </Grid>
-                                                <Grid item xs={2} sx={{ textAlign: 'start' }}>
-                                                    {`$ ${getUsersAmt(infoDict.user)}`}
+                                                <Grid item xs={4} sx={{ textAlign: 'end' }}>
+                                                    <strong>${calculatedTotalText}</strong>
                                                 </Grid>
-                                                <Grid item xs={2} sx={{ textAlign: 'end' }}>
-                                                    <TextField
-                                                        id={`split-value-${infoDict.user?.id}`}
-                                                        size='small'
-                                                        variant='standard'
-                                                        type='number'
-                                                        value={infoDict.value || ''}
-                                                        onChange={handleTextValueChange}
-                                                    />
-                                                </Grid>
-                                            </Grid>
-                                        ))}
-                                    </Grid>
-                                    <Grid item xs={12} sx={{ mt: 2 }}>
-                                        <Divider />
-                                        <Grid container sx={{ mt: 1 }}>
-                                            <Grid item xs={8}>
-                                                <strong>Total:</strong>
-                                            </Grid>
-                                            <Grid item xs={4} sx={{ textAlign: 'end' }}>
-                                                <strong>${calculatedTotalText}</strong>
                                             </Grid>
                                         </Grid>
                                     </Grid>
                                 </Grid>
-                            </Grid>
-                        )}
-                    </StyledTabPanel>
-                </DialogContent>
-            </Paper>
-        </Grid>
+                            )}
+                        </StyledTabPanel>
+                    </DialogContent>
+                </Paper>
+            </Grid>
+        </Dialog>
     );
 };
