+++ front-end/src/components/expense/styles/SplitDetailsTabStyles.js
@@ -36,7 +36,7 @@ export function StyledTabPanel(props) {
             aria-labelledby={`simple-tab-${index}`}
             {...other}
         >
-            {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
+            {value === index && <Box sx={{ p: 0 }}>{children}</Box>}
         </div>
     );
 }
